packages/sunagent-app/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sunagent-app"
version = "1.0.0"
license = {file = "LICENSE-CODE"}
description = "sunagent app"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "autogen-core",
  "autogen-ext",
  "autogen-agentchat",
  "autogen",
  "flask[async]",
  "flask_cors",
  "quart",
  "tweepy",
  "async_lru",
  "apscheduler",
  "pytz",
  "google-genai",
]

[dependency-groups]
dev = [
]

[tool.hatch.version]
path = "src/sunagent_app/version.py"

[project.scripts]
sunagent_app = "sunagent_app.app:main"

[tool.ruff]
extend = "../../pyproject.toml"
include = ["src/**", "tests/*.py"]

[tool.pyright]
extends = "../../pyproject.toml"
include = ["tests"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]

[tool.poe]
include = "../../shared_tasks.toml"

[tool.poe.tasks]
test = "pytest -n auto"
